Move bloc to bloc folder
