+++ lib/src/feature/initialization/logic/app_runner.dart
@@ -4,8 +4,8 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:logger/logger.dart';
-import 'package:sizzle_starter/src/core/common/app_bloc_observer.dart';
-import 'package:sizzle_starter/src/core/common/bloc_transformer.dart';
+import 'package:sizzle_starter/src/core/common/bloc/app_bloc_observer.dart';
+import 'package:sizzle_starter/src/core/common/bloc/bloc_transformer.dart';
 import 'package:sizzle_starter/src/core/common/error_reporter/error_reporter.dart';
 import 'package:sizzle_starter/src/core/constant/application_config.dart';
 import 'package:sizzle_starter/src/feature/initialization/logic/composition_root.dart';
